+++ .pre-commit-config.yaml
@@ -11,17 +11,6 @@ repos:
         # We use `--no-cache` to ensure that we always run based on what the
         # pre-commit tool provides us.
         args: [ --no-cache ]
-  - repo: https://github.com/pycqa/pylint
-    rev: v2.14.5
-    hooks:
-      - id: pylint
-  - repo: https://github.com/pycqa/flake8
-    rev: 6.0.0
-    hooks:
-      - id: flake8
-        args:
-          - --max-line-length=120
-          - --ignore=ANN101,ANN102,ANN204 # Ignoring annotation for self, cls, and special methods
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.4.0
     hooks:
