+++ tecton_client/_internal/async_tecton_client.py
@@ -1,20 +1,22 @@
 import json
-from typing import Optional, Dict, Union
+from typing import Dict
+from typing import Optional
+from typing import Union
 from urllib.parse import urljoin
 
 import httpx
 
-from tecton_client import TectonClient
-from tecton_client._internal.data_types import GetFeaturesResponse, GetFeatureServiceMetadataResponse
+from tecton_client._internal.data_types import GetFeatureServiceMetadataResponse
+from tecton_client._internal.data_types import GetFeaturesResponse
+from tecton_client._internal.tecton_client_base import TectonClientBase
+from tecton_client.exceptions import convert_exception
 
 
-class AsyncTectonClient(TectonClient):
+class AsyncTectonClient(TectonClientBase):
     # redefining init just to make the type of 'client' explicit
-    def __init__(self,
-                 url: str,
-                 api_key: str,
-                 default_workspace_name: Optional[str] = None,
-                 client: httpx.AsyncClient = None):
+    def __init__(
+        self, url: str, api_key: str, default_workspace_name: Optional[str] = None, client: httpx.AsyncClient = None
+    ):
         """Constructor for the client
 
         Args:
@@ -27,33 +29,56 @@ class AsyncTectonClient(TectonClient):
         """
         super().__init__(url=url, api_key=api_key, default_workspace_name=default_workspace_name, client=client)
 
-    # overwriting to make async
-    async def get_features(self,
-                           *,
-                           feature_service_name: Optional[str] = None,
-                           feature_service_id: Optional[str] = None,
-                           join_key_map: Optional[Dict[str, Union[int, str, type(None)]]] = None,
-                           request_context_map: Optional[Dict[str, Union[int, str, float]]] = None,
-                           metadata_options: Optional[Dict[str, bool]] = None,
-                           workspace_name: Optional[str] = None,
-                           request_options: Optional[Dict[str, bool]] = None,
-                           allow_partial_results: bool = False,
-                           ) -> GetFeaturesResponse:
-        return super().get_features(feature_service_name=feature_service_name,
-                                    feature_service_id=feature_service_id,
-                                    join_key_map=join_key_map,
-                                    request_context_map=request_context_map,
-                                    metadata_options=metadata_options,
-                                    workspace_name=workspace_name,
-                                    request_options=request_options,
-                                    allow_partial_results=allow_partial_results)
+    async def get_features(
+        self,
+        *,
+        feature_service_name: Optional[str] = None,
+        feature_service_id: Optional[str] = None,
+        join_key_map: Optional[Dict[str, Union[int, str, type(None)]]] = None,
+        request_context_map: Optional[Dict[str, Union[int, str, float]]] = None,
+        metadata_options: Optional[Dict[str, bool]] = None,
+        workspace_name: Optional[str] = None,
+        request_options: Optional[Dict[str, bool]] = None,
+        allow_partial_results: bool = False,
+    ) -> GetFeaturesResponse:
+        self._validate(feature_service_id, feature_service_name, workspace_name)
+        request_data = self._build_get_features_request(
+            feature_service_id,
+            feature_service_name,
+            join_key_map,
+            request_context_map,
+            workspace_name,
+            metadata_options,
+            allow_partial_results,
+            request_options,
+        )
+        resp = self._client.post(url=urljoin(self._base_url, "feature-service/get-features"), json=request_data)
+
+        try:
+            resp.raise_for_status()
+        except httpx.HTTPStatusError as exc:
+            raise convert_exception(exc) from exc
+
+        return GetFeaturesResponse.from_dict(json.loads(resp.text))
 
     # overwriting to make async
-    async def get_feature_service_metadata(self, *,
-                                           feature_service_name: Optional[str] = None,
-                                           feature_service_id: Optional[str] = None,
-                                           workspace_name: Optional[str] = None,
-                                           ) -> GetFeatureServiceMetadataResponse:
-        return super().get_feature_service_metadata(feature_service_name=feature_service_name,
-                                                    feature_service_id=feature_service_id,
-                                                    workspace_name=workspace_name)
+    async def get_feature_service_metadata(
+        self,
+        *,
+        feature_service_name: Optional[str] = None,
+        feature_service_id: Optional[str] = None,
+        workspace_name: Optional[str] = None,
+    ) -> GetFeatureServiceMetadataResponse:
+        self._validate(feature_service_id, feature_service_name, workspace_name)
+        request_data = self._build_get_feature_service_metadata_request(
+            feature_service_id=feature_service_id,
+            feature_service_name=feature_service_name,
+            workspace_name=workspace_name,
+        )
+        resp = self._client.post(url=urljoin(self._base_url, "feature-service/metadata"), json=request_data)
+        try:
+            resp.raise_for_status()
+        except httpx.HTTPStatusError as exc:
+            raise convert_exception(exc) from exc
+
+        return GetFeatureServiceMetadataResponse.from_dict(json.loads(resp.text))
