+++ tecton_client/_internal/tecton_client.py
@@ -8,11 +8,8 @@ from urllib.parse import urljoin
 import httpx
 from httpx import HTTPStatusError
 
-from tecton_client.__about__ import __version__ as tecton_version
 from tecton_client._internal.data_types import GetFeatureServiceMetadataResponse
 from tecton_client._internal.data_types import GetFeaturesResponse
-from tecton_client._internal.data_types import MetadataOptions
-from tecton_client._internal.data_types import RequestOptions
 from tecton_client.exceptions import convert_exception
 
 
@@ -32,20 +29,7 @@ class TectonClient:
             client: An httpx.Client, allowing you to provide finer-grained customization on the request behavior,
                 such as default timeout or connection settings. See https://www.python-httpx.org/ for more info.
         """
-        self.url = url
-        self.default_workspace_name = default_workspace_name
-        self._api_key = api_key
-        self._base_url = urljoin(url, "/api/v1/")
-
-        headers = httpx.Headers(
-            {"Authorization": "Tecton-key " + api_key, "User-Agent": "tecton-http-python-client " + tecton_version}
-        )
-        if client is None:
-            self._client = httpx.Client(headers=headers)
-        else:
-            self._client = client
-            # add the headers to the existing client headers
-            self._client.headers.update(headers)
+        super().__init__(url=url, api_key=api_key, default_workspace_name=default_workspace_name, client=client)
 
     def get_features(
         self,
@@ -99,62 +83,3 @@ class TectonClient:
             raise convert_exception(exc) from exc
 
         return GetFeatureServiceMetadataResponse.from_dict(json.loads(resp.text))
-
-    def _build_get_features_request(
-        self,
-        feature_service_id,
-        feature_service_name,
-        join_key_map,
-        request_context_map,
-        workspace_name,
-        metadata_options,
-        allow_partial_results,
-        request_options,
-    ):
-        workspace_name = workspace_name or self.default_workspace_name
-        params = {
-            "workspaceName": workspace_name,
-            "featureServiceName": feature_service_name,
-            "featureServiceId": feature_service_id,
-            "joinKeyMap": join_key_map or {},
-            "requestContextMap": request_context_map or {},
-            "allowPartialResults": allow_partial_results,
-        }
-        if metadata_options:
-            params["metadataOptions"] = {
-                # these two default to True
-                "includeNames": metadata_options.get(MetadataOptions.include_names, True),
-                "includeDataTypes": metadata_options.get(MetadataOptions.include_data_types, True),
-                # the rest default to False
-                "includeEffectiveTimes": metadata_options.get(MetadataOptions.include_effective_times, False),
-                "includeSloInfo": metadata_options.get(MetadataOptions.include_slo_info, False),
-                "includeServingStatus": metadata_options.get(MetadataOptions.include_serving_status, False),
-            }
-        if request_options:
-            params["requestOptions"] = {
-                "readFromCache": request_options.get(RequestOptions.read_from_cache),
-                "writeToCache": request_options.get(RequestOptions.write_to_cache),
-            }
-        request_data = {"params": params}
-        return request_data
-
-    def _build_get_feature_service_metadata_request(self, feature_service_id, feature_service_name, workspace_name):
-        return {
-            "params": {
-                "featureServiceName": feature_service_name,
-                "featureServiceId": feature_service_id,
-                "workspaceName": workspace_name,
-            }
-        }
-
-    def _validate(self, feature_service_id, feature_service_name, workspace_name):
-        workspace_name = workspace_name or self.default_workspace_name
-        if not workspace_name and not self.default_workspace_name:
-            msg = (
-                "workspace_name not set. Parameter workspace_name must be set either in TectonClient "
-                "initialization or in get_features"
-            )
-            raise ValueError(msg)
-        if not (feature_service_id or feature_service_name) or (feature_service_id and feature_service_name):
-            msg = "must pass exactly one of feature_service_name or feature_service_id"
-            raise ValueError(msg)
